重构：拆分插件加载与注册职责
